TETRIS-91: Input fields are weirdly arranged as well as the corresponding error messages (#12)
Fixed the alignment of input fields with the ugly "AddInputStyleOnLoad" script because the other option would have been "styling" the form in Python which sounded not very promising. Also, I added the "registry.js" script back to the template so the errors get rendered. The errors had their margin removed so it looks better.