Tetris 122: Implement lobby creation and invitation of people into a lobby (#20)
* TETRIS-129: Friends in the friends list disappear if the view changes

Introduces a method to get the friends list if it exists. It needs to be called on every view which extends tetris_app/base.html.

* TETRIS-122: Implement lobby creation

Lobbies kind of work but actually they don't lol

* TETRIS-122: Implement lobby creation

I absolutely know what I am doing

* TETRIS-122: Implement lobby creation

Changed it so that there is a single lobby id that is not regenerated on every page load.

* TETRIS-122: Implement lobby creation

I guess it works! (or at least I hope so). You can join a lobby now and invite other people.